+++ src/Core/TundraCore/Asset/IAssetTransfer.cpp
@@ -71,7 +71,7 @@ QByteArray IAssetTransfer::RawData() const
     if (rawAssetData.size() == 0) 
         return QByteArray(); 
     else 
-        return QByteArray((const char*)&rawAssetData[0], rawAssetData.size());
+        return QByteArray((const char*)&rawAssetData[0], (int)rawAssetData.size());
 }
 
 QString IAssetTransfer::SourceUrl() const
