+++ src/Core/TundraCore/Audio/OggVorbisLoader.cpp
@@ -20,7 +20,7 @@ namespace
 class OggMemDataSource
 {
 public:
-    OggMemDataSource(const u8* data, uint size) :
+    OggMemDataSource(const u8* data, size_t size) :
         data_(data),
         size_(size),
         position_(0)
@@ -29,8 +29,9 @@ public:
 
     size_t Read(void* ptr, size_t size)
     {
-        uint max_read = size_ - position_;
-        if (size > max_read) size = max_read;
+        size_t max_read = size_ - position_;
+        if (size > max_read)
+            size = max_read;
         if (size)
         {
             memcpy(ptr, &data_[position_], size);
@@ -70,8 +71,8 @@ public:
         
 private:
     const u8* data_;
-    uint size_;
-    uint position_;        
+    size_t size_;
+    size_t position_;
 };
 
 size_t OggReadCallback(void* ptr, size_t size, size_t nmemb, void* datasource)
